+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file follows the format suggested by [Keep a CHANGELOG](https://github.com/olivierlacan/keep-a-changelog).
 
 ## [Unreleased][Unreleased]
+
+## [17.3.0][17.3.0] - 2016-09-16
 ### Added
 - [Feature] Add `isActive` prop to Button React Component
 
@@ -565,7 +567,7 @@ This file follows the format suggested by [Keep a CHANGELOG](https://github.com/
 ### Added
 - [Patch] Use semantic versioning in LEGO (#58).
 
-[Unreleased]: https://github.com/optimizely/oui/compare/v17.2.0...HEAD
+[Unreleased]: https://github.com/optimizely/oui/compare/v17.3.0...HEAD
 [0.0.2]: https://github.com/optimizely/oui/compare/v0.0.1...v0.0.2
 [0.0.3]: https://github.com/optimizely/oui/compare/v0.0.2...v0.0.3
 [1.0.0]: https://github.com/optimizely/oui/compare/v0.0.3...v1.0.0
@@ -618,3 +620,4 @@ This file follows the format suggested by [Keep a CHANGELOG](https://github.com/
 [17.0.0]: https://github.com/optimizely/oui/compare/v16.0.0...v17.0.0
 [17.1.0]: https://github.com/optimizely/oui/compare/v17.0.0...v17.1.0
 [17.2.0]: https://github.com/optimizely/oui/compare/v17.1.0...v17.2.0
+[17.3.0]: https://github.com/optimizely/oui/compare/v17.2.0...v17.3.0